requirements.txt
opencv-python
matplotlib
numpy
Pillow
pretrainedmodels
scikit-learn
scipy
segmentation-models-pytorch
tensorflow
tensorflow-estimator
torch
torchvision
tqdm